requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
apscheduler==3.10.4
markdown2==2.4.10